+++ evm/src/main/java/org/hyperledger/besu/evm/operation/AbstractCreateOperation.java
@@ -92,7 +92,7 @@ public abstract class AbstractCreateOperation extends AbstractOperation {
     final Wei value = Wei.wrap(frame.getStackItem(0));
 
     final Address address = frame.getRecipientAddress();
-    final MutableAccount account = frame.getWorldUpdater().getAccount(address);
+    final MutableAccount account = getMutableAccount(address, frame);
 
     frame.clearReturnData();
 
@@ -184,19 +184,25 @@ public abstract class AbstractCreateOperation extends AbstractOperation {
     parent.decrementRemainingGas(childGasStipend);
 
     // frame addition is automatically handled by parent messageFrameStack
-    MessageFrame.builder()
-        .parentMessageFrame(parent)
-        .type(MessageFrame.Type.CONTRACT_CREATION)
-        .initialGas(childGasStipend)
-        .address(contractAddress)
-        .contract(contractAddress)
-        .inputData(inputData)
-        .sender(parent.getRecipientAddress())
-        .value(value)
-        .apparentValue(value)
-        .code(code)
-        .completer(child -> complete(parent, child, evm))
-        .build();
+    MessageFrame.Builder builder =
+        MessageFrame.builder()
+            .parentMessageFrame(parent)
+            .type(MessageFrame.Type.CONTRACT_CREATION)
+            .initialGas(childGasStipend)
+            .address(contractAddress)
+            .contract(contractAddress)
+            .inputData(inputData)
+            .sender(parent.getRecipientAddress())
+            .value(value)
+            .apparentValue(value)
+            .code(code)
+            .completer(child -> complete(parent, child, evm));
+
+    if (parent.getEip7928AccessList().isPresent()) {
+      builder.eip7928AccessList(parent.getEip7928AccessList().get());
+    }
+
+    builder.build();
 
     parent.setState(MessageFrame.State.CODE_SUSPENDED);
   }