+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ### Additions and Improvements
 - Update log4j [#9131](https://github.com/hyperledger/besu/pull/9131)
 - Expose new method to query hardfork by block number Plugin API [#9115](https://github.com/hyperledger/besu/pull/9115)
+- Support loading multiple transaction selector plugins [#8743](https://github.com/hyperledger/besu/pull/9139)
 
 #### Performance
 - Add jmh benchmarks for some compute-related opcodes [#9069](https://github.com/hyperledger/besu/pull/9069)